# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Students API",
        "description": "API for managing students",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/students": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Get all students",
                "description": "Display a listing of the resource.",
                "operationId": "944b18b2938371d3a49103c865f5aed9",
                "responses": {
                    "200": {
                        "description": "List of all students",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Student"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Students"
                ],
                "summary": "Create a new student",
                "description": "Store a newly created resource in storage.",
                "operationId": "55c75624de9d3fe47487bf48df409809",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StudentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Student created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Student"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/students/{id}": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Get a student by ID",
                "description": "Display the specified resource.",
                "operationId": "3d03685fc4e68bbf325c181306b4a037",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Student ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Student details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Student"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Student not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Students"
                ],
                "summary": "Update a student",
                "description": "Update the specified resource in storage.",
                "operationId": "c29610976f0d6bce915b31aee0e85a2f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Student ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StudentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Student not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Students"
                ],
                "summary": "Delete a student",
                "description": "Remove the specified resource from storage.",
                "operationId": "d96ee6ae75daa51d0106f3146837dc86",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Student ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Student not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Student": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "FirstName": {
                        "type": "string",
                        "example": "John"
                    },
                    "LastName": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "School": {
                        "type": "string",
                        "example": "BCIT"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "StudentRequest": {
                "required": [
                    "FirstName",
                    "LastName",
                    "School"
                ],
                "properties": {
                    "FirstName": {
                        "type": "string",
                        "example": "John"
                    },
                    "LastName": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "School": {
                        "type": "string",
                        "example": "BCIT"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Students",
            "description": "Student management endpoints"
        }
    ]
}